Wait for pid to disappear and tidy
