+++ .gitignore
@@ -13,5 +13,6 @@ runIOC.sh
 relPaths.sh
 *.py[oc]
 *.sq3
-/.idea
+**/.idea
 .project
+*.py.bak
